Fixed header 'RELO' fied
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@8053 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/p2p/MessageStructsGen.hxx
@@ -156,7 +156,7 @@ class ForwardingHeaderStruct : public PDU {
 public:
    ForwardingHeaderStruct();
    ForwardingHeaderStruct(const ForwardingHeaderStruct&);
-   UInt8                         mReloToken;
+   UInt32                        mReloToken;
    UInt32                        mOverlay;
    UInt8                         mTtl;
    UInt8                         mReserved;
